requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dataclasses==0.8
Flask==2.0.1
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.0
MarkupSafe==2.0.1
PyJWT==1.7.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
twilio==6.59.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
